refactor: split rumors data layer into rumors_data.go
